# Add --no-location option to the run command