--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -31,6 +31,8 @@ Released 2023-07-19
 - Added a simple profiler to the probe-rs cli toolkit (#1628)
 - Added MSP432E4 target (MSP432E401Y and MSP432E411Y). (#1139)
 - probe-rs-cli: added `attach` subcommand. (#1672, #1616)
+- probe-rs-cli: added --no-locations option to the run command, which suppresses the filename and line number
+                information from the rtt log (#1704)
 
 ### Changed