@Yatekii 2023-07-28
Please format the code with rustfmt =)