@elfmimi 2023-07-29
noticed that change log entry for this pr has been put under 0.20.0. @Yatekii